+++ QUANTAXIS/QAUtil/QAParameter.py
@@ -100,11 +100,11 @@ class ORDER_STATUS():
 class AMOUNT_MODEL():
     """订单的成交量
 
-    by_price是按固定成交总额下单,动态计算成交量
+    by_money是按固定成交总额下单,动态计算成交量
     by_amount 按固定成家量下单
     """
 
-    BY_PRICE = 'by_price'
+    BY_MONEY = 'by_money'
     BY_AMOUNT = 'by_amount'